# internal/surface/http/openapi.json
{
  "components": {
    "parameters": {
      "Id": {
        "description": "The resource's ID, or (for an account or category) its unique name.",
        "in": "path",
        "name": "id",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "InvalidInput": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        },
        "description": "The request failed validation."
      },
      "NotFound": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        },
        "description": "No such resource."
      },
...
    },
    "schemas": {
...
      "CreateTransactionRequest": {
        "properties": {
          "account": {
            "description": "An account's ID or unique name.",
            "type": "string"
          },
          "amount": {
            "description": "Always positive; type says which way the money moves.",
            "format": "money",
            "type": "string"
          },
          "category": {
            "description": "A category's ID or unique name.",
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "date": {
            "description": "Omit to book the transaction to today in the account owner's own timezone.",
            "format": "date",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "type": {
            "enum": [
              "outflow",
              "inflow"
            ],
            "type": "string"
          }
        },
        "required": [
          "type",
          "account",
          "amount",
          "description"
        ],
        "type": "object"
      },
...
      "EditTransactionRequest": {
        "properties": {
          "account": {
            "description": "An account's ID or unique name. Read when editing an outflow or an inflow.",
            "type": "string"
          },
          "amount": {
            "description": "Always positive; the transaction's own type says which way the money moves.",
            "format": "money",
            "type": "string"
          },
          "category": {
            "description": "A category's ID or unique name. Read when editing an outflow or an inflow.",
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "date": {
            "description": "Omit to book the transaction to today in the account owner's own timezone.",
            "format": "date",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "from_account": {
            "description": "An account's ID or unique name. Read when editing a transfer.",
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "to_account": {
            "description": "An account's ID or unique name. Read when editing a transfer.",
            "type": "string"
          }
        },
        "required": [
          "amount",
          "description"
        ],
        "type": "object"
      },
      "Error": {
        "properties": {
          "code": {
            "enum": [
              "conflict",
              "internal",
              "invalid_input",
              "not_allowed",
              "not_found",
              "precondition_failed",
              "unauthenticated",
              "unavailable"
            ],
            "type": "string"
          },
          "details": {
            "additionalProperties": {},
            "description": "Safe, structured extra context - candidate names, valid options.",
            "type": "object"
          },
          "field": {
            "description": "The request field or CLI flag this error attaches to, when there is one.",
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "Transaction": {
        "properties": {
          "account_id": {
            "description": "Set on an outflow or an inflow.",
            "type": "string"
          },
          "amount": {
            "description": "Always positive; the transaction's type says which way the money moved.",
            "format": "money",
            "type": "string"
          },
          "category_id": {
            "description": "Set on an outflow or an inflow that has a category.",
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "date": {
            "description": "The date the transaction is booked to.",
            "format": "date",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "from_account_id": {
            "description": "Set on a transfer: the account the money left.",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "to_account_id": {
            "description": "Set on a transfer: the account the money arrived in.",
            "type": "string"
          },
          "type": {
            "enum": [
              "outflow",
              "inflow",
              "transfer"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "date",
          "description",
          "amount",
          "currency"
        ],
        "type": "object"
      },
      "TransactionEnvelope": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Transaction"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
      "TransactionList": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Transaction"
            },
            "type": "array"
          },
          "next_cursor": {
            "description": "An opaque token for the next page. Present only when there is a next page; its encoding isn't part of the API's contract and may change.",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
      "TransactionListEnvelope": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TransactionList"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "description": "bodger's read/write REST API. This is the web UI's only interface (milestone 2 generates its TypeScript client from this description) and is available to external clients and integrations. Every route but /healthz and POST /api/v1/auth/login requires a session cookie (from login) or an Authorization: Bearer API token; a cookie-authenticated state-changing request also needs the X-Bodger-CSRF header. The server binds a loopback address only unless a password has been set. Monetary values are always JSON strings, never numbers.",
    "title": "bodger REST API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/transactions": {
      "get": {
        "operationId": "listTransactions",
        "parameters": [
          {
            "description": "An account's ID or unique name.",
            "in": "query",
            "name": "account",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "A category's ID or unique name; its whole subtree is included.",
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "One of \"outflow\", \"inflow\", or \"transfer\".",
            "in": "query",
            "name": "type",
            "schema": {
              "enum": [
                "outflow",
                "inflow",
                "transfer"
              ],
              "type": "string"
            }
          },
          {
            "description": "The inclusive start of a booked-date range.",
            "in": "query",
            "name": "from",
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "description": "The inclusive end of a booked-date range.",
            "in": "query",
            "name": "to",
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "description": "Page size; defaults to 50 and caps at 200.",
            "in": "query",
            "name": "limit",
            "schema": {
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "An opaque token from a previous page's \"next_cursor\". Its encoding is not part of the API's contract and may change.",
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionListEnvelope"
                }
              }
            },
            "description": "A page of matching transactions."
          },
          "422": {
            "$ref": "#/components/responses/InvalidInput"
          }
        },
        "summary": "List transactions, newest first, with an optional filter."
      },
      "post": {
        "description": "For a transfer between two of the actor's own accounts, use POST /api/v1/transfers instead. When \"date\" is omitted, the server books the transaction to today in the actor's own timezone - never the client's.",
        "operationId": "createTransaction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionEnvelope"
                }
              }
            },
            "description": "The recorded transaction."
          },
          "422": {
            "$ref": "#/components/responses/InvalidInput"
          }
        },
        "summary": "Record an outflow or an inflow."
      }
    },
    "/api/v1/transactions/{id}": {
      "delete": {
        "description": "The transaction stops appearing in lists and balances, but its history is retained.",
        "operationId": "deleteTransaction",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionEnvelope"
                }
              }
            },
            "description": "The deleted transaction."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "Delete a transaction."
      },
      "get": {
        "operationId": "getTransaction",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionEnvelope"
                }
              }
            },
            "description": "The transaction."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "Fetch one transaction by ID."
      },
      "patch": {
        "description": "This is a full replacement of every field in the request body, not a partial patch: send every field's current value, not only the one that's changing. A transaction's type (outflow, inflow, transfer) can't change.",
        "operationId": "editTransaction",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditTransactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionEnvelope"
                }
              }
            },
            "description": "The updated transaction."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/InvalidInput"
          }
        },
        "summary": "Replace a transaction's editable fields."
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}